+++ Makefile
@@ -9,8 +9,8 @@ CFLAGS += -Wmissing-declarations
 $(BUILD)/db_tests: hiredis test/db_tests.c thirdparty/greatest.h event_loop.c state/redis.c common.c
 	$(CC) -o $@ test/db_tests.c event_loop.c state/redis.c common.c thirdparty/hiredis/libhiredis.a $(CFLAGS) -I. -Ithirdparty
 
-$(BUILD)/event_loop_tests: test/event_loop_tests.c thirdparty/greatest.h event_loop.c
-	$(CC) -o $@ test/event_loop_tests.c event_loop.c $(CFLAGS) -I. -Ithirdparty
+$(BUILD)/socket_tests: test/socket_tests.c thirdparty/greatest.h sockets.c
+	$(CC) -o $@ test/socket_tests.c sockets.c $(CFLAGS) -I. -Ithirdparty
 
 clean:
 	rm -r $(BUILD)/*
@@ -23,6 +23,6 @@ hiredis:
 
 test: hiredis redis $(BUILD)/db_tests $(BUILD)/event_loop_tests FORCE
 	./thirdparty/redis-3.2.3/src/redis-server &
-	sleep 1s ; ./build/db_tests ; ./build/event_loop_tests
+	sleep 1s ; ./build/db_tests ; ./build/socket_tests
 
 FORCE:
